+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.7.0] - unreleased
+### Added
+- `RuleGroup::limit()`
+- `RuleGroup::last_evaluation()`
+- `RuleGroup::evaluation_time()`
 ### Changed
 - Refactored deserialization of Prometheus server responses so the explicit dependency on `serde_json` could be removed.
 - Refactored the `error` module and some custom errors related to deserialization using `serde::de::Error`. The `Error` enum inside the `error` module now contains one variant less and existing error variants were improved by _properly_ implementing `std::error::Error::source()`. Libraries like `anyhow` are now able to display more detailed error messages. This change is not breaking if you did not match on specific error enum variants before.